Split out `wildcard_enum_match_arm` and `match_wildcard_for_single_variants`
